+++ CHANGELOG.md
@@ -2,21 +2,30 @@
 
 All versions below are listed in reverse chronological order.
 
+## [0.3.0](https://github.com/Heiko-san/curses_fzf/releases/tag/0.3.0) (2026-03-08)
+
+### Added
+
+- Added parameter `min_items` to FuzzyFinder, to specify the minimum number of selected
+  items to return. If the number is below this threshold, ENTER is not accepted.
+- Added parameter `max_items` to FuzzyFinder, to specify the maximum number of selected
+  items to return. If the number is above this threshold, ENTER is not accepted.
+
 ## [0.2.5](https://github.com/Heiko-san/curses_fzf/releases/tag/0.2.5) (2026-03-07)
 
 ### Changed
 
-- Better logic for scoring_fzf.
+- Better logic for `scoring_fzf`.
 
 ## [0.2.4](https://github.com/Heiko-san/curses_fzf/releases/tag/0.2.4) (2026-03-06)
 
 ### Added
 
-- Added scoring_fzf scoring function.
+- Added `scoring_fzf` scoring function.
 
 ### Changed
 
-- Changed default scoring function to scoring_fzf.
+- Changed default scoring function to `scoring_fzf`.
 
 ## [0.2.3](https://github.com/Heiko-san/curses_fzf/releases/tag/0.2.3) (2026-03-05)